fix(ADA-1785) - Change NavigationItem structure (#386)
This solves this https://kaltura.atlassian.net/browse/ADA-1785 and this
https://kaltura.atlassian.net/browse/ADA-2116
This issue was created by nesting ExpandedText that is a button inside
NavigationItem that is also a button. I could not use ExpandedText no
more, because I had to have the "More" button outside NaviationItem's
container. To fix this I used conditional rendering. To keep the same
styling I had to use position absolute for the "More" Button.